+++ src/components/app-tabs/Hints.vue
@@ -1,13 +1,15 @@
 <script setup lang="ts">
-import { computed, onActivated } from 'vue';
+import { computed, onActivated, ref, useTemplateRef, watchEffect } from 'vue';
 import check from '@/assets/icons/check.png';
 import minus from '@/assets/icons/minus.png';
 import { showMouseToast, ui } from '@/state/ui';
-import { hints, loadHints, type LocalHint } from '@/state/hints';
+import { getCssVarFromStatus, getHintStatusName, hints, hintsLastUpdated, HintStatus, loadHints, type LocalHint } from '@/state/hints';
 import AppTable, { type Column } from '@/components/AppTable.vue';
 import { settings } from '@/state/settings';
 import PlayerName from '../text-elements/PlayerName.vue';
 import ItemName from '../text-elements/ItemName.vue';
+import { client } from '@/lib/archipelago';
+import { useElementBounding, useEventListener } from '@vueuse/core';
 
 onActivated(async () => {
   loadHints();
@@ -18,7 +20,8 @@ const columns: Column[] = [
   { label: 'Player', key: 'player' },
   { label: 'Item', key: 'item' },
   { label: 'Owner', key: 'owner' },
-  { label: 'Location', key: 'location' }
+  { label: 'Location', key: 'location' },
+  { label: 'Status', key: 'status' }
 ];
 
 const computedColumns = computed(() => {
@@ -27,13 +30,16 @@ const computedColumns = computed(() => {
   }
 
   return columns;
-})
+});
 
 const filteredHints = computed(() => {
-  if (settings.value.hintsFilterFound) {
-    return hints.list.filter(hint => !hint.found);
-  }
-  return hints.list;
+  return hints.list.filter(hint => {
+    const isFound = hint.found;
+
+    if (!settings.value.hintsFilterFound) return true;
+
+    return !isFound;
+  });
 });
 
 async function copyHint(item: LocalHint) {
@@ -61,10 +67,60 @@ async function copyHint(item: LocalHint) {
     showMouseToast('Hint copied to clipboard');
   }
 }
+
+const priorityDropDown = useTemplateRef('priority-drop-down');
+const priorityDropDownBounding = useElementBounding(priorityDropDown);
+const changingPriority = ref(false);
+const priorityChangerTarget = ref<HTMLElement | null>(null);
+const priorityChangingHint = ref<LocalHint | null>(null);
+
+const priorityChangerStyle = computed(() => {
+  if (!priorityChangerTarget.value) return null;
+
+  const { width } = priorityDropDownBounding;
+  const targetBounding = priorityChangerTarget.value.getBoundingClientRect();
+
+  return {
+    top: `${targetBounding.top}px`,
+    left: `${targetBounding.left - width.value - 4}px`,
+    pointerEvents: (priorityDropDownBounding.width.value === 0 ? 'none' : 'auto') as any
+  };
+});
+
+useEventListener(['click', 'wheel', 'contextmenu', 'pointerdown'], () => {
+  changingPriority.value = false;
+  priorityChangerTarget.value = null;
+});
+
+function showPriorityDropDown(e: PointerEvent, item: LocalHint) {
+  const target = e.target as HTMLElement;
+  priorityChangerTarget.value = target;
+  changingPriority.value = true;
+  priorityChangingHint.value = item;
+}
+
+function changePriority(newStatus: HintStatus) {
+  if (!priorityChangingHint.value) return;
+
+  client.socket.send({
+    cmd: 'UpdateHint' as any,
+    location: priorityChangingHint.value.locationId,
+    player: priorityChangingHint.value.ownerSlot,
+    status: newStatus
+  });
+}
 </script>
 
 <template>
   <div class="hints">
+    <Transition>
+      <div ref="priority-drop-down" v-if="changingPriority && priorityChangerStyle" :style="priorityChangerStyle" class="priority-drop-down">
+        <div @click="changePriority(HintStatus.NO_PRIORITY)" :style="{ color: getCssVarFromStatus(HintStatus.NO_PRIORITY) }">No Priority</div>
+        <div @click="changePriority(HintStatus.PRIORITY)" :style="{ color: getCssVarFromStatus(HintStatus.PRIORITY) }">Priority</div>
+        <div @click="changePriority(HintStatus.AVOID)" :style="{ color: getCssVarFromStatus(HintStatus.AVOID) }">Avoid</div>
+      </div>
+    </Transition>
+
     <div class="actions">
       <button :disabled="hints.points < hints.cost" @click="ui.modals.buyItemHint = true">Buy item hint</button>
       <div class="check-row" style="margin: 0.5em 0;">
@@ -105,6 +161,17 @@ async function copyHint(item: LocalHint) {
         <template #location="{ item }">
           <td style="color: var(--theme-location)">{{ item.location }}</td>
         </template>
+
+        <template #status="{ item }">
+          <td :style="{ color: getCssVarFromStatus(item.status) + ' !important' }">
+            <template v-if="!item.found && item.player === client.players.self.alias">
+              <a @click.prevent.stop="e => showPriorityDropDown(e, item)" :style="{ color: getCssVarFromStatus(item.status) + ' !important' }" class="priority-changer">{{ getHintStatusName(item.status) }}</a>
+            </template>
+            <template v-else>
+              {{ getHintStatusName(item.status) }}
+            </template>
+          </td>
+        </template>
       </AppTable>
     </div>
   </div>
@@ -140,5 +207,38 @@ async function copyHint(item: LocalHint) {
 .copy-btn svg {
   width: 1.2em;
   height: 1.2em;
+}
+
+.priority-changer {
+  text-decoration: underline;
+  cursor: context-menu;
+}
+
+.priority-drop-down {
+  position: fixed;
+  background: white;
+  padding: 0.2em;
+  border-radius: 4px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+  z-index: 99999999;
+}
+
+.priority-drop-down > div {
+  padding: 0.5em 1em;
+  user-select: none;
+  cursor: pointer;
+  border-radius: 2px;
+}
+
+.priority-drop-down > div:hover {
+  background: rgba(0, 0, 0, 0.1);
+}
+
+.v-enter-active, .v-leave-active {
+  transition: opacity 0.1s ease;
+}
+
+.v-enter-from, .v-leave-to {
+  opacity: 0;
 }
 </style>