+++ src/state/changelog.ts
@@ -5,8 +5,6 @@ import { ref } from 'vue';
 export const changelogHTML = ref('');
 
 async function parseChangelog() {
-  console.log('parse')
-
   const result = await marked.parse(changelog.replace(/^[\u200B\u200C\u200D\u200E\u200F\uFEFF]/,''), {
     async: true,
     gfm: true